added phones to mobilePhonesPage and create Card
