+++ src/App.scss
@@ -1,7 +1,7 @@
 @import "./styles/reset";
 @import "./styles/utils/mixins";
 @import "./styles/utils/vars";
-// @import "./styles/typography";
+@import "./styles/typography";
 
 * {
   box-sizing: border-box;
