+++ src/components/Card/Card.scss
@@ -1,0 +1,112 @@
+.wrap {
+  margin-right: 16px;
+  background: #fff;
+  border: 1px solid #e2e6e9;
+  box-sizing: border-box;
+  border-radius: 5px;
+}
+
+.card__container-inner {
+  margin: 24px;
+}
+
+.card {
+  box-sizing: border-box;
+  width: 270px;
+
+  &__img {
+    width: 208px;
+    height: 208px;
+    margin: 32px 32px 0;
+
+  }
+
+  &__title {
+    // font-family: Mont;
+    height: 28px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 14px;
+    line-height: 21px;
+    color: #313237;
+  }
+
+  &__prise {
+    // font-family: Mont;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 22px;
+    line-height: 140%;
+    color: #313237;
+  }
+
+  &__oldPrise {
+    // font-family: Mont;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 22px;
+    line-height: 28px;
+    text-decoration-line: line-through;
+    color: #89939a;
+  }
+
+  &__info {
+    border-top: 1px solid #e2e6e9;
+  }
+
+  &__item {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 8px;
+  }
+
+  &__info-screen {
+    margin-top: 16px;
+  }
+
+  &__info-screen_name {
+    // font-family: Mont;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 12px;
+    line-height: 15px;
+    color: #89939a;
+  }
+
+  &__info-screen_value {
+    // font-family: Mont;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 12px;
+    line-height: 15px;
+    text-align: right;
+    color: #313237;
+  }
+
+  &__button-wrap {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  &__button-cart {
+    height: 40px;
+    width: 176px;
+    border: none;
+    background: #313237;
+    color: #fff;
+    transition: 0.3s;
+    cursor: pointer;
+
+    &:hover {
+      background: #fff;
+      color: #27ae60;
+      border: 1px solid #e2e6e9;
+    }
+  }
+
+  &__button-favor {
+    height: 40px;
+    width: 40px;
+    border: none;
+  }
+}
