+++ src/components/ProductsSlider/ProductsSlider.tsx
@@ -1,19 +1,7 @@
 import React, { useState } from 'react';
 import './ProductsSlider.scss';
-// import { getProducts } from '../../helpers/api';
-// import { getHotProducts } from '../../helpers/api';
+import { Card } from '../Card/Card';
 
-interface Slide {
-  imageUrl: string;
-  name: string;
-  snippet: string;
-  price: number;
-  discount: number;
-  screen: string;
-  capacity: string;
-  ram: string;
-  id: string;
-}
 type ProductsSliderProps = {
   title: string;
   visibleProducts: Slide[];
@@ -56,7 +44,7 @@ const ProductsSlider: React.FC<ProductsSliderProps> = ({ visibleProducts, title 
             className="Carousel__list"
           >
             {visibleProducts.map(product => (
-              <Item key={product.id} {...product} />
+              <Card key={product.id} {...product} />
             ))}
           </div>
         </div>
@@ -70,75 +58,5 @@ const ProductsSlider: React.FC<ProductsSliderProps> = ({ visibleProducts, title 
     </div>
   );
 };
-
-type Props = {
-  imageUrl: string;
-  name: string;
-  price: number;
-  discount: number;
-  screen: string;
-  capacity: string;
-  ram: string;
-};
-
-const Item: React.FC<Props> = ({
-  imageUrl,
-  name,
-  price,
-  discount,
-  screen,
-  capacity,
-  ram,
-}) => (
-  <div className="wrap">
-    <article className="card">
-      <img alt="card" src={imageUrl} className="card__img" />
-      <div className="card__container-inner">
-        <h3 className="card__title">{name}</h3>
-        <span className="card__prise">
-          $
-          {(price - price * (discount / 100))}
-        </span>
-        {' '}
-        {discount !== 0
-         && (
-           <span className="card__oldPrise">
-             $
-             {price}
-           </span>
-         )}
-
-        <div className="card__info">
-          <div className="card__info-screen card__item">
-            <p className="card__info-screen_name">Screen</p>
-            <p className="card__info-screen_value">{screen}</p>
-          </div>
-          <div className="card__info-capacity card__item">
-            <p className="card__info-screen_name">Capacity</p>
-            <p className="card__info-screen_value">{capacity}</p>
-          </div>
-          <div className="card__info-ram card__item">
-            <p className="card__info-screen_name">RAM</p>
-            <p className="card__info-screen_value">{ram}</p>
-          </div>
-        </div>
-        <div className="card__button-wrap">
-          <button
-            type="button"
-            className="card__button-cart"
-          >
-            Add to cart
-          </button>
-          <button
-            type="button"
-            className="card__button-favor"
-          >
-            favor
-          </button>
-        </div>
-      </div>
-    </article>
-  </div>
-);
 
 export default ProductsSlider;
