+++ src/styles/_typography.scss
@@ -2,19 +2,19 @@
   font-family: "Mont-Bold", sans-serif;
   font-style: normal;
   font-weight: bold;
-  src: url("./fonts/Mont-Bold.otf");
+  src: url("../../public/fonts/Mont-Bold.otf");
 }
 
 @font-face {
   font-family: "Mont-Reguler", sans-serif;
   font-style: normal;
   font-weight: 500;
-  src: url("./fonts/Mont-Reguler.otf");
+  src: url("../../public/fonts/Mont-Regular.otf");
 }
 
 @font-face {
   font-family: "Mont-SemiBold", sans-serif;
   font-style: normal;
   font-weight: 600;
-  src: url("./fonts/Mont-SemiBold.otf");
+  src: url("../../public/fonts/Mont-SemiBold.otf");
 }
